+++ tests/entrypoints/openai/responses/test_protocol.py
