+++ tests/entrypoints/openai/responses/test_responses_utils.py
