+++ tests/entrypoints/serve/disagg/test_serving_tokens.py
@@ -12,7 +12,7 @@ from vllm.config import ModelConfig
 from vllm.config.utils import getattr_iter
 from vllm.v1.engine.detokenizer import check_stop_strings
 
-from ...utils import RemoteOpenAIServer
+from ....utils import RemoteOpenAIServer
 
 MODEL_NAME = "Qwen/Qwen3-0.6B"
 GEN_ENDPOINT = "/inference/v1/generate"
