+++ tests/multimodal/test_embedding_shape_validation.py
